two assets for off pieces for each color
off_<color>_shadow: 200 x 66 incl a small shadow that the checker would have if a little space were between it and the next one.

off_<color>_last: 200x56 without the shadow. Either use only these and add 10 space to get your normal 200x66 spacing, or use n-1 times the one with the shadow and then one without it at the end where a shadow would overlap the container in a way that it doesn't.