Fixed checker look
If there are multiple checker assets in a design, then now what asset is used for a checker depends solely on its number on its point, ie. the 3rd checker on point 14 will always look the same. This way all checkers remain the same before and after moves, except for the checker that has moved.

Every point now gets an independent but constant "random start index" for determining the right asset for each checker.